requirements.txt
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-ckeditor==6.1.0
django-etc==1.3.0
django-filebrowser==3.14.1
django-grappelli==2.15.1
django-hitcount==1.3.3
django-js-asset==1.2.2
dnspython==2.1.0
filelock==3.3.1
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
idna
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
